pyproject.toml
[project]
name = "llm-rag-chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "killermama98",email = "ctkiena2@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10.0, <3.13"
dependencies = [
    "numpy (>=1.2.2,<3.0.0)",
    "langchain (>=0.3.16,<0.4.0)",
    "streamlit (>=1.41.1,<2.0.0)",
    "langchain-openai (>=0.3.2,<0.4.0)",
    "langchain-milvus (>=0.1.8,<0.2.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain-community (>=0.3.16,<0.4.0)",
    "rank-bm25 (>=0.2.2,<0.3.0)",
    "beautifulsoup4 (>=4.12.3,<5.0.0)",
    "lxml (>=5.3.0,<6.0.0)",
    "unstructured[all-docs] (==0.16.6)",
    "unstructured-pytesseract (>=0.3.13,<0.4.0)",
    "langchain-weaviate (>=0.0.4,<0.0.5)",
    "langchain-mongodb (>=0.4.0,<0.5.0)",
    "pymongo (>=4.11,<5.0)",
    "streamlit-feedback (>=0.1.4,<0.2.0)",
    "duckduckgo-search (>=7.3.2,<8.0.0)"
]

[tool.poetry]
packages = [{include = "src"}]


[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
